libccp: update libccp for new API
This commit updates libccp to follow the new CCP API, which
gets rid of control patterns and fold functions and combines them
into one program that the user sends. The program contains
events, which are expressions that evaluate to a boolean; with each
event, there is a set of instructions to run if that condition is true.
This commit also updates libccp to implement new types of registers,
local and implicit registers, and separates the scope of user defined
variables into control and report registers. Now, libccp reads only
expression messages and update fields messages. The new state machine code
is in machine.c. Finally, this commit adds unit tests
that test the new API in test.c