libccp: deduplicate programs across flows
This commit changes libccp to store datapath programs in memory
independent of single flows, to reduce memory usage. It introduces
seriaization for a new message, change program, that changes the current
datapath program being run for that flow; all programs should be sent
down at the beginning in an install message.