ccp_invoke: Retransmit create messages
The send_conn_create function should not modify the sent_create
variable, since it could fail. Rather, the caller (ccp_connection_start
in the common case, and ccp_invoke if a retransmission is necessary)
should set this bit.

This had caused a bug where the create message was erroneously believed
to have been sent to CCP sucessfully when it had not been.