integration_test: Script to run CCP integration test
https://github.com/ccp-project/portus already implements an integration
test for interoperability between portus and libccp. However, proposed
modifications to libccp are not currently tested in CI for compatibility
with existing portus master branch. This script fills the gap by cloning
the portus repository, checking out the current libccp commit in the
libccp submodule within the integration test, and running the
integration test.