remove all global state
Prior to this commit, libccp stored all internal state in two global
data structures: a struct `ccp_datapath`, and an array of struct
`ccp_connection`s.

However, newer userspace datapaths, such as mvfst and mtcp, internally
use multiple threads, each tied to a core, and each handling their own
separate set of connections. Using libccp with global state would
require (unnecessary) synchronization, because technically they are not
touching any of the same state.

This commit slightly changes the API to remove all global state:

1. `ccp_init` still allocates space for the struct ccp_datapath, but
instead of storing the pointer internally in a global variable, it
returns it to the caller. The caller is now responsible for storing this
pointer somewhere and including it in all future calls to libccp
functions.
2. The `ccp_active_connections` array is now stored within the
`ccp_datapath` structure in order to tie them together
3. The `ccp_connection` structure now maintains a pointer to its parent
datapath. This prevents having to always pass around the pair of
ccp_connetion and ccp_datapath since they are almost always both needed
when dealing with a connection.